Object extraction not working

+++ lerf/data/lerf_datamanager.py
@@ -43,7 +43,7 @@ from nerfstudio.data.datamanagers.base_datamanager import VanillaDataManager, Va
 @dataclass
 class LERFDataManagerConfig(VanillaDataManagerConfig):
     _target: Type = field(default_factory=lambda: LERFDataManager)
-    patch_tile_size_range: Tuple[int, int] = (0.05, 0.5)
+    patch_tile_size_range: Tuple[float, float] = (0.05, 0.5)
     patch_tile_size_res: int = 7
     patch_stride_scaler: float = 0.5
